pyproject.toml
[project]
name = "solana-verifiable-lottery"
version = "1.0.0"
description = "Verifiable Solana token-holder lottery using blockhash seed and reproducible entrant snapshot."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "base58>=2.1.1",
  "python-dotenv>=1.0.1",
]

[project.scripts]
solana-lottery = "solana_lottery.cli:main"

[tool.ruff]
line-length = 100


[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
